Add delete game swipe

+++ composeApp/src/commonMain/kotlin/com/ionvaranita/belotenote/datalayer/database/dao/groups2/Game2GroupsDao.kt
@@ -21,11 +21,11 @@ interface Game2GroupsDao {
     suspend fun insert(game2GroupsEntity: Game2GroupsEntity): Long
 
     @Query("update Game2GroupsEntity set statusGame = :statusGame, winnerPoints = :winnerPoints, dateGame = :dateGame where idGame = :idGame")
-    suspend fun update(idGame: Short, statusGame: Byte, winnerPoints: Short, dateGame: Long) : Int
+    suspend fun update(idGame: Int, statusGame: Byte, winnerPoints: Short, dateGame: Long) : Int
 
     @Query("update Game2GroupsEntity set statusGame = :statusGame, dateGame = :dateGame where idGame = :idGame")
-    suspend fun updateStatus(idGame: Short, statusGame: Byte, dateGame: Long): Int
+    suspend fun updateStatus(idGame: Int, statusGame: Byte, dateGame: Long): Int
 
     @Query("delete from Game2GroupsEntity  where idGame = :idGame")
-    suspend fun delete(idGame: Short)
+    suspend fun delete(idGame: Int)
 }
