+++ composeApp/src/commonMain/kotlin/com/ionvaranita/belotenote/datalayer/database/dao/players3/Game3PDao.kt
@@ -15,14 +15,14 @@ interface Game3PDao {
     fun getGames(): Flow<List<Game3PEntity>>
 
     @Query("select * from Game3PEntity where idGame = :idGame")
-    suspend fun getGame(idGame: Short): Game3PEntity
+    suspend fun getGame(idGame: Int): Game3PEntity
 
     @Query("update Game3PEntity set statusGame = :statusGame, dateGame = :dateGame where idGame = :idGame")
-    suspend fun updateStatus(idGame: Short, statusGame: Byte, dateGame: Long): Int
+    suspend fun updateStatus(idGame: Int, statusGame: Byte, dateGame: Long): Int
 
     @Query("update Game3PEntity set statusGame = :statusGame, winnerPoints = :winnerPoints, dateGame = :dateGame where idGame = :idGame")
-    suspend fun update(idGame: Short, statusGame: Byte, winnerPoints: Short, dateGame: Long): Int
+    suspend fun update(idGame: Int, statusGame: Byte, winnerPoints: Short, dateGame: Long): Int
 
     @Query("delete from Game3PEntity where idGame = :idGame")
-    suspend fun delete(idGame: Short)
+    suspend fun delete(idGame: Int)
 }
