+++ composeApp/src/commonMain/kotlin/com/ionvaranita/belotenote/datalayer/database/dao/players4/Game4PDao.kt
@@ -18,14 +18,14 @@ interface Game4PDao {
     fun getGames(): Flow<List<Game4PEntity>>
 
     @Query("select * from Game4PEntity where idGame = :idGame")
-    suspend fun getGame(idGame: Short): Game4PEntity
+    suspend fun getGame(idGame: Int): Game4PEntity
 
     @Query("update Game4PEntity set statusGame = :statusGame, dateGame = :dateGame where idGame = :idGame")
-    suspend fun updateStatus(idGame: Short, statusGame: Byte, dateGame: Long): Int
+    suspend fun updateStatus(idGame: Int, statusGame: Byte, dateGame: Long): Int
 
     @Query("update Game4PEntity set statusGame = :statusGame, winnerPoints = :winnerPoints, dateGame = :dateGame where idGame = :idGame")
-    suspend fun update(idGame: Short, statusGame: Byte, winnerPoints: Short, dateGame: Long): Int
+    suspend fun update(idGame: Int, statusGame: Byte, winnerPoints: Short, dateGame: Long): Int
 
     @Query("delete from Game4PEntity where idGame = :idGame")
-    suspend fun delete(idGame: Short)
+    suspend fun delete(idGame: Int)
 }
