+++ composeApp/src/commonMain/kotlin/com/ionvaranita/belotenote/datalayer/database/entity/groups2/Game2GroupsEntity.kt
@@ -11,7 +11,7 @@ import com.ionvaranita.belotenote.constants.GameStatus
  */
 @Entity
 data class Game2GroupsEntity(@PrimaryKey(
-    autoGenerate = true) val idGame: Short = 0, val dateGame: Long = getCurrentTime(), val statusGame: Byte = GameStatus.CONTINUE.id, val winnerPoints: Short, val name1: String, val name2: String) {
+    autoGenerate = true) val idGame: Int = 0, val dateGame: Long = getCurrentTime(), val statusGame: Byte = GameStatus.CONTINUE.id, val winnerPoints: Short, val name1: String, val name2: String) {
     fun toUiModel(): Game2GroupsUi {
         return Game2GroupsUi(idGame = this.idGame, statusGame = this.statusGame, winnerPoints = this.winnerPoints, name1 = this.name1, name2 = this.name2)
     }
