+++ composeApp/src/commonMain/kotlin/com/ionvaranita/belotenote/datalayer/database/entity/players4/Game4PEntity.kt
@@ -11,7 +11,7 @@ import com.ionvaranita.belotenote.constants.GameStatus
  */
 @Entity
 data class Game4PEntity(@PrimaryKey(
-    autoGenerate = true) val idGame: Short = 0, val dateGame: Long = getCurrentTime(), val statusGame: Byte = GameStatus.CONTINUE.id, val winnerPoints: Short, val name1: String, val name2: String, val name3: String, val name4: String) {
+    autoGenerate = true) val idGame: Int = 0, val dateGame: Long = getCurrentTime(), val statusGame: Byte = GameStatus.CONTINUE.id, val winnerPoints: Short, val name1: String, val name2: String, val name3: String, val name4: String) {
     fun toUiModel(): Game4PUi {
         return Game4PUi(idGame = this.idGame, statusGame = this.statusGame, winnerPoints = this.winnerPoints, name1 = this.name1, name2 = this.name2, name3 = this.name3, name4 = this.name4)
     }
